old-tests/submission/Infraware/Forms/contents/Forms/input_formmethod.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_formMethod</h3>
	</p>
	
	<hr>

	<div id="log"></div> 

	<form method="get"
		  enctype=""
		  action=""
		  name="input_form">
	<p><input type='hidden' id='input_text'></p>
	</form>

	<script>
		
		var input_text = document.getElementById("input_text");
		input_text.formMethod = "post";

		if (typeof(input_text.formMethod) == "string") {
			test(function() {
				assert_equals(input_text.formMethod, "post", "formEnctype attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formMethod attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
